BM-8[feat]: 경매 서비스 인터페이스 구현체 분리
